[spantree] move neighbors fn to link_graph
