+++ src/algorithm/spantree.rs
@@ -26,7 +26,7 @@ pub fn span(
     while !queue.is_empty() {
         let v = queue.pop_front().unwrap();
         visited.insert(v.clone());
-        for n in neighbors(g, &v) {
+        for n in g.neighbors(&v) {
             if visited.insert(n.clone()) {
                 queue.push_back(n.clone());
                 result.insert(n, v.clone());
@@ -36,36 +36,9 @@ pub fn span(
     result
 }
 
-fn neighbors(
-    g: &impl GraphDCEL<
-        LinkVertex,
-        LinkDart,
-        LinkFace,
-        LinkGraphIter<LinkVertex>,
-        LinkGraphIter<LinkDart>,
-        LinkGraphIter<LinkFace>,
-    >,
-    v: &LinkVertex,
-) -> Vec<LinkVertex> {
-    let mut current_dart = g.dart_vertex(v);
-    let first_dart = current_dart.clone();
-    let mut current_neighbor = g.dart_target(&current_dart);
-    let mut result = vec![];
-    loop {
-        result.push(current_neighbor);
-        let twin_dart = g.twin(&current_dart);
-        current_dart = g.next(&twin_dart);
-        if current_dart == first_dart {
-            break;
-        }
-        current_neighbor = g.dart_target(&current_dart);
-    }
-    result
-}
-
 #[cfg(test)]
 mod tests {
-    use crate::algorithm::spantree::{neighbors, span};
+    use crate::algorithm::spantree::span;
     use crate::data_structure::link_graph::LinkGraph;
     use std::collections::HashMap;
 
@@ -155,76 +128,5 @@ mod tests {
         assert_eq!(edges.len(), 2);
         assert_eq!(edges.get(&lv2), Some(&lv1));
         assert_eq!(edges.get(&lv0), Some(&lv1));
-    }
-
-    #[test]
-    fn neighbors_single_edge() {
-        let mut lg = LinkGraph::new();
-        let lv1 = lg.new_vertex();
-        let lv2 = lg.new_vertex();
-
-        let ld1 = lg.new_dart(lv1.clone(), lv2.clone(), None, None, None, None);
-        let lf = lg.new_face(ld1.clone());
-        lg.new_dart(
-            lv2.clone(),
-            lv1.clone(),
-            Some(ld1.clone()),
-            Some(ld1.clone()),
-            Some(ld1),
-            Some(lf),
-        );
-
-        assert_eq!(neighbors(&lg, &lv1), vec![lv2.clone()]);
-        assert_eq!(neighbors(&lg, &lv2), vec![lv1]);
-    }
-
-    #[test]
-    fn neighbors_triangle() {
-        let mut lg = LinkGraph::new();
-        let lv0 = lg.new_vertex();
-        let lv1 = lg.new_vertex();
-        let lv2 = lg.new_vertex();
-
-        let ld0 = lg.new_dart(lv0.clone(), lv1.clone(), None, None, None, None);
-        let lf = lg.new_face(ld0.clone());
-        let ld1 = lg.new_dart(
-            lv1.clone(),
-            lv2.clone(),
-            Some(ld0.clone()),
-            None,
-            None,
-            Some(lf.clone()),
-        );
-        let ld2 = lg.new_dart(
-            lv2.clone(),
-            lv0.clone(),
-            Some(ld1.clone()),
-            Some(ld0.clone()),
-            None,
-            Some(lf),
-        );
-
-        let lt0 = lg.new_dart(lv1.clone(), lv0.clone(), None, None, Some(ld0), None);
-        let lof = lg.new_face(lt0.clone());
-        let lt2 = lg.new_dart(
-            lv0.clone(),
-            lv2.clone(),
-            Some(lt0.clone()),
-            None,
-            Some(ld2),
-            Some(lof.clone()),
-        );
-        lg.new_dart(
-            lv2.clone(),
-            lv1.clone(),
-            Some(lt2),
-            Some(lt0),
-            Some(ld1),
-            Some(lof),
-        );
-
-        assert_eq!(neighbors(&lg, &lv0), vec![lv2.clone(), lv1.clone()]);
-        assert_eq!(neighbors(&lg, &lv1), vec![lv0.clone(), lv2.clone()]);
-        assert_eq!(neighbors(&lg, &lv2), vec![lv1, lv0]);
     }
 }
