+++ src/data_structure/link_graph.rs
@@ -203,6 +203,23 @@ impl
         LinkGraphIter::new(self.faces.clone())
     }
 
+    fn neighbors(&self, vertex: &LinkVertex) -> Vec<LinkVertex> {
+        let mut current_dart = self.dart_vertex(vertex);
+        let first_dart = current_dart.clone();
+        let mut current_neighbor = self.dart_target(&current_dart);
+        let mut result = vec![];
+        loop {
+            result.push(current_neighbor);
+            let twin_dart = self.twin(&current_dart);
+            current_dart = self.next(&twin_dart);
+            if current_dart == first_dart {
+                break;
+            }
+            current_neighbor = self.dart_target(&current_dart);
+        }
+        result
+    }
+
     fn dart_vertex(&self, vertex: &LinkVertex) -> LinkDart {
         vertex.0.borrow().dart.clone().unwrap()
     }
@@ -457,5 +474,76 @@ mod tests {
         let f2 = fi.next().unwrap();
         assert_eq!(f1, f1);
         assert_ne!(f1, f2);
+    }
+
+    #[test]
+    fn neighbors_single_edge() {
+        let mut lg = LinkGraph::new();
+        let lv1 = lg.new_vertex();
+        let lv2 = lg.new_vertex();
+
+        let ld1 = lg.new_dart(lv1.clone(), lv2.clone(), None, None, None, None);
+        let lf = lg.new_face(ld1.clone());
+        lg.new_dart(
+            lv2.clone(),
+            lv1.clone(),
+            Some(ld1.clone()),
+            Some(ld1.clone()),
+            Some(ld1),
+            Some(lf),
+        );
+
+        assert_eq!(lg.neighbors(&lv1), vec![lv2.clone()]);
+        assert_eq!(lg.neighbors(&lv2), vec![lv1]);
+    }
+
+    #[test]
+    fn neighbors_triangle() {
+        let mut lg = LinkGraph::new();
+        let lv0 = lg.new_vertex();
+        let lv1 = lg.new_vertex();
+        let lv2 = lg.new_vertex();
+
+        let ld0 = lg.new_dart(lv0.clone(), lv1.clone(), None, None, None, None);
+        let lf = lg.new_face(ld0.clone());
+        let ld1 = lg.new_dart(
+            lv1.clone(),
+            lv2.clone(),
+            Some(ld0.clone()),
+            None,
+            None,
+            Some(lf.clone()),
+        );
+        let ld2 = lg.new_dart(
+            lv2.clone(),
+            lv0.clone(),
+            Some(ld1.clone()),
+            Some(ld0.clone()),
+            None,
+            Some(lf),
+        );
+
+        let lt0 = lg.new_dart(lv1.clone(), lv0.clone(), None, None, Some(ld0), None);
+        let lof = lg.new_face(lt0.clone());
+        let lt2 = lg.new_dart(
+            lv0.clone(),
+            lv2.clone(),
+            Some(lt0.clone()),
+            None,
+            Some(ld2),
+            Some(lof.clone()),
+        );
+        lg.new_dart(
+            lv2.clone(),
+            lv1.clone(),
+            Some(lt2),
+            Some(lt0),
+            Some(ld1),
+            Some(lof),
+        );
+
+        assert_eq!(lg.neighbors(&lv0), vec![lv2.clone(), lv1.clone()]);
+        assert_eq!(lg.neighbors(&lv1), vec![lv0.clone(), lv2.clone()]);
+        assert_eq!(lg.neighbors(&lv2), vec![lv1, lv0]);
     }
 }
